requirements.txt
langchain-community
openai
langchain-core
langgraph
langchain
langchain-openai
pydantic
IPython
dotenv
requests
beautifulsoup4
python-docxs